Move marker chart to More menu
Move the marker breakdown chart button under the 'More...' section options menu,
make it a static class, and rename it to MarkerBreakdownChart.

Since the section options header isn't available when extended marker stats
aren't enable, keep the button where it is in that case.

Also fix a bug in Overlay code that could result in overlays fighting for
control if:
1. An overlay is already showing
2. We request to show another overlay, triggering an animated dismissal of the
   old overlay
3. Before that dismissal completes, another overlay request comes in.