Support marker references in time input
Initial support for more complex references in time inputs, though more work is
needed to support them in bulk operations, where they'd be the most useful.

All this logic is contained in a new TimeInput class, which also encapsulates
(and adapts) the existing timeInputShortcutHandler logic.

In addition to standard hh:mm:ss.000 and ms notation, allow markers to be
referenced when the input starts with an '='. In this mode, the following
expression can contain:
  * 0 or 1 marker expression of the form MI[L], where
    * M: marker type. M for any, I for intro, C for credits, A for ad
    * I: marker index. Positive or negative. M3 is the third marker, C-1 is the
      last credits marker.
    * L: location. S for the marker's start timestamp, E for end. Will be
      inferred if not present.
  * 0 or more standard hms or ms time expressions, joined with '+' or '-'.
    There's not really a reason to have multiple, but M2+1+2 is valid.
  * The marker expression can't be subtracted: 1000-M1 is not valid.
  * If a marker expression is present, the timestamp cannot go negative.

Due to overlap, in "complex" mode, the "big" timestamp modifier shortcuts have
been changed to 'o' and 'p' instead of '-' and '='. I need to consider whether
everything should switch over, since it's not the best user experience to have
different shortcuts based on what's in the field.

While complex expressions aren't supported in bulk operations, switch over to
TimeInputs for bulk add and shift so the old shortcut handlers can be ripped out
entirely.

Minimal testing, but all the core operations should work as expected.

Tangential changes:
* Add title to text input for complex/non-hms expressions to show hms.
* Add minify parameter to msToHms to leave out unnecessary minutes/thousandths.
* Attach paste listener to all time inputs, not just MarkerTable inputs.
* Background color changes when the time input cannot be parsed.