Add experimental ability to write extended data
Currently markers are only written to the taggings table, which does everything
we need it to, but there are other references to these markers in the extra_data
column of the media_parts table. Add the ability to adjust that field as well
with updated data.

The downside is that this extra_data is another point of failure - if Plex
changes the layout of this field, more might go wrong than having some bad
standalone entries in the taggings table.

This change also adds a somewhat hacky json-only setting to adjust Express's
trust proxy setting, as a way to work around potential issues with rate limiting
and reverse proxies.