Tentative fix for #38
I had some hacky logic that queries the database differently if some queries
involved more than 500 IDs (as a way to get around SQLite's conditional operator
limitations). In that 500+ case, there's a bug when querying episode IDs that
would prevent the operation from succeeding if the first episode did not have
any markers associated with it. Rework the query building logic to remove the
need for separate logic when looking up large sets of data.

Details:
Previously I was creating a chain of "WHERE x=? OR x=? OR x=? OR ...", which
could run into SQLite 3 limitations (e.g. exceeding SQLITE_MAX_VARIABLE_NUMBER).
Rewriting these to instead use "IN", and directly adding the values to the query
instead of using parameters (after ensuring we actually have integers) seems to
appease SQLite on my machine, even when there are 1K+ items to query. It's
likely not very efficient, but it's a relatively uncommon scenario that doesn't
need to be super optimized.