Add option to suspend server after inactivity
Add an option to automatically disconnect from the Plex database after a certain
time of inactivity. Marker Editor assumes that Plex's database does not change
after it starts, but that's not necessarily the case if Plex is running
alongside this application (which isn't recommended, but still works on some
systems). This setting allows Marker Editor to detach from Plex's database after
a period of inactivity, and automatically reconnect once a new request comes in,
allowing Marker Editor to load in any new markers that have been added.

This has only been lightly tested, so it's likely bugs still exist, but the core
ability to disconnect/reattach to the database works as expected.